refactor: change tomcat scope and move WebServerLauncher to main
