+++ week5/mawakeb/src/main/java/next/webserver/RequestMapping.java
@@ -1,4 +1,4 @@
-package next.servlet;
+package next.webserver;
 
 import next.controller.common.Controller;
 import next.controller.HomeController;
